+++ Source/Core/Log/NovusLogger.Provider.Files.pas
@@ -15,7 +15,7 @@ type
     fbDailyRotate: Boolean;
     fsFilename: String;
     fiLimitLogSize : Int64;
-    procedure WriteLine(aLogMessage: String);
+    procedure WriteFile(aLogMessage: String);
   Public
     constructor Create(aFilename: String;
                        aDailyRotate: Boolean = false;
@@ -138,7 +138,7 @@ begin
   AddLog(aLogMessage, Now, TSeverityType.stException);
 end;
 
-procedure TNovusLogger_Provider_Files.WriteLine(aLogMessage: String);
+procedure TNovusLogger_Provider_Files.WriteFile(aLogMessage: String);
 Var
   FStreamWriter: tStreamWriter;
   liRetryCount: Integer;
@@ -151,7 +151,7 @@ begin
       FStreamWriter := TStreamWriter.Create(fsfilename,true);
       FStreamWriter.AutoFlush := true;
 
-      FStreamWriter.WriteLine(aLogMessage);
+      FStreamWriter.WriteFile(aLogMessage);
 
       FStreamWriter.Flush;
 
@@ -183,7 +183,7 @@ end;
 
 procedure TNovusLogger_Provider_Files.SendLogMessage(aLogMessage: String; aLogDateTime: tDateTime; aSeverityType: TSeverityType);
 begin
-  WriteLine(aLogMessage);
+  WriteFile(aLogMessage);
 end;